Refactor type definition
* Split across multiple files.

* Rename types, most noticeably `TestInterface` is now `TestFn`, but
really anything that had the Interface suffix has been changed.

* Update some documentation.
