+++ test-d/context.ts
@@ -1,12 +1,12 @@
 import {expectError, expectType} from 'tsd';
 
-import anyTest, {ExecutionContext, TestInterface} from '..';
+import anyTest, {ExecutionContext, TestFn} from '..';
 
 interface Context {
 	foo: string;
 }
 
-const test = anyTest as TestInterface<Context>; // eslint-disable-line @typescript-eslint/no-unnecessary-type-assertion
+const test = anyTest as TestFn<Context>; // eslint-disable-line @typescript-eslint/no-unnecessary-type-assertion
 
 const macro = test.macro((t, expected: number) => {
 	expectType<string>(t.context.foo);
@@ -27,6 +27,6 @@ interface Covariant extends Context {
 	bar: number;
 }
 
-const test2 = anyTest as TestInterface<Covariant>; // eslint-disable-line @typescript-eslint/no-unnecessary-type-assertion
+const test2 = anyTest as TestFn<Covariant>; // eslint-disable-line @typescript-eslint/no-unnecessary-type-assertion
 const hook = (t: ExecutionContext<Context>) => {};
 test2.beforeEach(hook);
