+++ types/subscribable.ts
@@ -1,0 +1,6 @@
+export interface Subscribable {
+	subscribe(observer: {
+		error(error: any): void;
+		complete(): void;
+	}): void;
+}
